not working, untested
going to sleep for the night, but wanted to commit this
basic code is in, looks okay, unrun, uncompiled even,
certainly untested.

modified:   lib/DB/CouchDB.pm

I added an attachment function, hacked _call to detect whether
an attachment is being done (got a little schizo on the options there,
and need to go back and fix that), and make a new url type for attachments

Still need to test, of course, and also want to add calls to delete
attachments while I'm at it.