fixed deletion
two fixes here.  First, allow doc to be passed as is to delete.
Delete will now take either a doc name and rev pair, or just
the doc object with the embedded _id and _rev.

Second, I uri_escaped the id, so that long complicated ids with
slashes in them still work.